server/pyproject.toml
[tool.poetry]
name = "easyctf-platform"
version = "0.1.0"
description = ""
authors = ["Michael Zhang <mail@mzhang.io>"]
license = "AGPL-3.0"

[tool.poetry.dependencies]
python = "^3.10"
bcrypt = "^4.0.1"
celery = "^5.2.7"
coverage = "^6.5.0"
cryptography = "^38.0.4"
Flask = "^2.2.2"
Flask-Breadcrumbs = "^0.5.1"
Flask-Caching = "^2.0.1"
Flask-Celery-Helper = "^1.1.0"
flask-csp = "^0.10"
Flask-Login = "^0.6.2"
Flask-Migrate = "^4.0.0"
Flask-Script = "^2.0.6"
flask-sqlalchemy = "^3.0.2"
Flask-WTF = "^1.0.1"
gitdb = "^4.0.10"
GitPython = "^3.1.29"
markdown2 = "^2.4.6"
mysqlclient = "^2.1.1"
onetimepass = "^1.0.1"
paramiko = "^2.12.0"
passlib = "^1.7.4"
pathlib = "^1.0.1"
Pillow = "^9.3.0"
pycryptodome = "^3.16.0"
PyQRCode = "^1.2.1"
pytest = "^7.2.0"
PyYAML = "^6.0"
rauth = "^0.7.3"
raven = {extras = ["flask"], version = "^6.10.0"}
redis = "^4.3.5"
requests = "^2.28.1"
SQLAlchemy = "^1.4.44"
WTForms-Components = "^0.10.5"
Werkzeug = "^2.2.2"
cachelib = "^0.9.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
